# warehouse/src/main/resources/swagger/warehouse.yml
openapi: 3.0.3
info:
  title: Warehouse
  description: Warehouse stores Product(s)
  version: 1.0.0
servers:
  - url: 'https://localhost:8081'
tags:
  - name: product
    description: warehouse handler
paths:
  "/wh/product/{id}":
    get:
      tags:
        - product
      summary: get one product by id from warehouse
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResultDto"

    delete:
      tags:
        - product
      summary: delete product from warehouse
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string

  "/wh/product":
    get:
      tags:
        - product
      summary: get all products in warehouse
      operationId: getAllProducts
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResultDto"

    post:
      tags:
        - product
      summary: add product in warehouse
      operationId: addProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPostDto"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResultDto"

    put:
      tags:
        - product
      summary: update product in warehouse
      operationId: updateProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateDto"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResultDto"



components:
  schemas:
    ProductResultDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: integer
          format: int32
        amount:
          type: integer
          format: int32
    ProductUpdateDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: integer
          format: int32
        amount:
          type: integer
          format: int32
    ProductPostDto:
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
          format: int32
        amount:
          type: integer
          format: int32